Make Account a hidden trigger
Since this isn’t a reliable way to tell when new accounts have been added to Flow, we don’t want to allow our users to poll this endpoint for "New Account created" logic, but we still need it in order to populate the task creation assignee field.

It is possible for there to be accounts in Flow that are not members of any workspaces, and the current way that we are fetching accounts is through the `memberships` endpoint, which only will include accounts that are members of workspaces.

The public API does not expose any better way to fetch accounts at the moment, however.