Make task output more user-friendly
Cherry pick the values from the API so it's not such a messy data dump, and define pretty labels for the values that better match the app's language.

One thing I'd love to do but Zapier's documentation is confusing me so it can be a "do-later" thing: Instead of exposing List (Project) ID, show the List's (Project's) name alongside the ID. Same thing with Owner, Account, etc.