Add updated_after param to task list request
Since zapier polls every 5-15 minutes we don't need ALL tasks, just
tasks updated within a recent time frame.

However, we only do this for the "real" request, as for sample data
there may not be any recently updated tasks but we still want to pull
something in to test out.