Add view: member param to workspaces endpoint
This isn't currently documented on our public API docs but is supported.
We don't want to get ALL workspaces because that includes workspaces an
account is not currently a member of but is allowed to join.

If we  get all those additional workspaces and use it to populate the
task creation form, the API will reply with an error if the user tries
to create a task in a workspace they are not actually a member of.